+++ contracts/modules/oracle/Oracle.sol
@@ -2,11 +2,12 @@
 pragma solidity ^0.8.9;
 
 import "chainlink/interfaces/FeedRegistryInterface.sol";
-import { Denominations } from "chainlink/Denominations.sol";
+import {Denominations} from "chainlink/Denominations.sol";
 import "../../interfaces/IOracle.sol";
 
 contract Oracle is IOracle {
-    FeedRegistryInterface internal registry; 
+    FeedRegistryInterface internal registry;
+
     constructor(address _registry) {
         registry = FeedRegistryInterface(_registry);
     }
@@ -14,17 +15,18 @@ contract Oracle is IOracle {
     /**
      * Returns the latest price in USD to 8 decimals
      */
-    function getLatestAnswer(address token) external returns (int) {
+    function getLatestAnswer(address token) external returns (int256) {
         (
-            /* uint80 roundID */, 
-            int price,
-            /* uint80 startedAt */,
-            /* uint80 timeStamp */,
+            /* uint80 roundID */
+            ,
+            int256 price,
+            /* uint80 startedAt */
+            ,
+            /* uint80 timeStamp */
+            ,
             /* uint80 answeredInRound */
         ) = registry.latestRoundData(token, Denominations.USD);
-        
+
         return price;
     }
-
-    
 }
